Move feature toggles test util to package
